Fix to make friendlier to multi-OS use
This should make things nicer for using by non-Linux POSIX complaiant
things.
